requirements.txt
# Backend Gateway
Flask
Flask-CORS
gunicorn

# Agent and Orchestration
langchain
langgraph
langchain-core
langchain-community
langchain-groq
pydantic

# Security and Guardrails
presidio-analyzer
presidio-anonymizer
spacy

# Sandbox
sympy
ddgs

# Observability
python-dotenv
langsmith

# Jupyter Notebook
ipykernel